+++ packages/create-analog/template-minimal/vite.config.ts
@@ -1,0 +1,23 @@
+/// <reference types="vitest" />
+
+import { defineConfig } from 'vite';
+import analog from '@analogjs/platform';
+
+// https://vitejs.dev/config/
+export default defineConfig(({ mode }) => ({
+  build: {
+    target: ['es2020'],
+  },
+  resolve: {
+    mainFields: ['module'],
+  },
+  plugins: [
+    analog({
+      ssr: false,
+      static: true,
+      prerender: {
+        routes: [],
+      },
+    }),
+  ],
+}));