+++ packages/vite-plugin-angular/src/lib/angular-build-optimizer-plugin.ts
@@ -1,68 +1,44 @@
 import { Plugin } from 'vite';
-import { transformAsync } from '@babel/core';
-import { createEs2015LinkerPlugin as linkerPluginCreator } from '@angular/compiler-cli/linker/babel';
 
-import { angularApplicationPreset, requiresLinking } from './utils/devkit.js';
+import { JavaScriptTransformer } from './utils/devkit.js';
 
 export function buildOptimizerPlugin({
   isProd,
-  supportedBrowsers,
 }: {
   isProd: boolean;
   supportedBrowsers: string[];
 }): Plugin {
+  const javascriptTransformer = new JavaScriptTransformer(
+    {
+      sourcemap: false,
+      thirdPartySourcemaps: false,
+      advancedOptimizations: true,
+      jit: true,
+    },
+    1
+  );
+
   return {
     name: '@analogjs/vite-plugin-angular-optimizer',
     apply: 'build',
-    config(userConfig) {
+    config() {
       return {
-        esbuild: !userConfig.esbuild
-          ? {
-              legalComments: 'none',
-              keepNames: false,
-              define: isProd
-                ? {
-                    ngDevMode: 'false',
-                    ngJitMode: 'false',
-                    ngI18nClosureMode: 'false',
-                  }
-                : undefined,
-              supported: {
-                // Native async/await is not supported with Zone.js. Disabling support here will cause
-                // esbuild to downlevel async/await to a Zone.js supported form.
-                'async-await': false,
-                // Zone.js also does not support async generators or async iterators. However, esbuild does
-                // not currently support downleveling either of them. Instead babel is used within the JS/TS
-                // loader to perform the downlevel transformation. They are both disabled here to allow
-                // esbuild to handle them in the future if support is ever added.
-                // NOTE: If esbuild adds support in the future, the babel support for these can be disabled.
-                'async-generator': false,
-                'for-await': false,
-                'class-field': false,
-                'class-static-field': false,
-              },
-            }
-          : {
-              define: isProd
-                ? {
-                    ngDevMode: 'false',
-                    ngJitMode: 'false',
-                    ngI18nClosureMode: 'false',
-                  }
-                : undefined,
-            },
+        esbuild: {
+          define: isProd
+            ? {
+                ngDevMode: 'false',
+                ngJitMode: 'false',
+                ngI18nClosureMode: 'false',
+              }
+            : undefined,
+        },
       };
     },
     async transform(code, id) {
       if (/\.[cm]?js$/.test(id)) {
-        const angularPackage = /[\\/]node_modules[\\/]@angular[\\/]/.test(id);
-        const forceAsyncTransformation =
-          !/[\\/][_f]?esm2015[\\/]/.test(id) &&
-          /for\s+await\s*\(|async\s+function\s*\*/.test(code);
-        const shouldLink = await requiresLinking(id, code);
-        const useInputSourcemap = (!isProd ? undefined : false) as undefined;
+        const angularPackage = /fesm20/.test(id);
 
-        if (!forceAsyncTransformation && !isProd && !shouldLink) {
+        if (!angularPackage) {
           return {
             code: isProd
               ? code.replace(/^\/\/# sourceMappingURL=[^\r\n]*/gm, '')
@@ -73,38 +49,14 @@ export function buildOptimizerPlugin({
           };
         }
 
-        const result = await transformAsync(code, {
-          filename: id,
-          inputSourceMap: useInputSourcemap,
-          sourceMaps: !isProd ? 'inline' : false,
-          compact: false,
-          configFile: false,
-          babelrc: false,
-          browserslistConfigFile: false,
-          plugins: [],
-          presets: [
-            [
-              angularApplicationPreset,
-              {
-                angularLinker: {
-                  shouldLink,
-                  jitMode: false,
-                  linkerPluginCreator,
-                },
-                forceAsyncTransformation,
-                supportedBrowsers,
-                optimize: isProd && {
-                  looseEnums: angularPackage,
-                  pureTopLevel: angularPackage,
-                },
-              },
-            ],
-          ],
-        });
+        const result: Uint8Array = await javascriptTransformer.transformData(
+          id,
+          code,
+          false
+        );
 
         return {
-          code: result?.code || '',
-          map: result?.map as any,
+          code: Buffer.from(result).toString(),
         };
       }